Refactor subscription query on frontend
