+++ frontend/src/app/can-ndjson-stream.d.ts
